+++ src/main/java/com/avab/avab/controller/AuthController.java
@@ -1,0 +1,28 @@
+package com.avab.avab.controller;
+
+import com.avab.avab.apiPayload.BaseResponse;
+import com.avab.avab.dto.response.AuthResponseDTO.OAuthResponse;
+import com.avab.avab.dto.response.AuthResponseDTO.TokenRefreshResponse;
+import com.avab.avab.service.impl.AuthServiceImpl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class AuthController {
+
+    private final AuthServiceImpl authServiceImpl;
+
+    @GetMapping("/auth/login/kakao")
+    public BaseResponse<OAuthResponse> kakaoLogin(@RequestParam("code") String code) {
+        return BaseResponse.onSuccess(authServiceImpl.kakaoLogin(code));
+    }
+
+    @PostMapping("/auth/refresh")
+    public BaseResponse<TokenRefreshResponse> refresh(@ExtractToken String refreshToken) {
+        return BaseResponse.onSuccess(authServiceImpl.refresh(refreshToken));
+    }
+}
