+++ src/main/java/com/avab/avab/dto/OAuthToken.java
@@ -1,9 +1,10 @@
-package com.avab.avab.security.kakao.dto;
+package com.avab.avab.dto;
 
 import lombok.Data;
 
 @Data
 public class OAuthToken {
+
     private String access_token;
     private String token_type;
     private String refresh_token;
