+++ src/main/java/com/avab/avab/dto/response/AuthResponseDTO.java
@@ -1,0 +1,32 @@
+package com.avab.avab.dto.response;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class AuthResponseDTO {
+
+    @Getter
+    @Builder
+    @AllArgsConstructor(access = AccessLevel.PROTECTED)
+    @NoArgsConstructor(access = AccessLevel.PROTECTED)
+    public static class OAuthResponse {
+
+        Long userId;
+        String accessToken;
+        String refreshToken;
+        Boolean isLogin;
+    }
+
+    @Getter
+    @Builder
+    @AllArgsConstructor(access = AccessLevel.PROTECTED)
+    @NoArgsConstructor(access = AccessLevel.PROTECTED)
+    public static class TokenRefreshResponse {
+
+        String accessToken;
+        String refreshToken;
+    }
+}
