+++ src/main/java/com/avab/avab/repository/UserRepository.java
@@ -1,10 +1,11 @@
 package com.avab.avab.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.avab.avab.domain.User;
+import com.avab.avab.domain.enums.SocialType;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    User findByEmail(String email);
+    Optional<User> findByEmailAndSocialType(String email, SocialType socialType);
 }
