--- src/main/java/com/avab/avab/security/kakao/dto/LoginResponse.java
@@ -1,13 +1,0 @@
-package com.avab.avab.security.kakao.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-@AllArgsConstructor
-public class LoginResponse {
-    private String accessToken;
-    private String refreshToken;
-}
