+++ src/main/java/com/avab/avab/service/AuthService.java
@@ -1,0 +1,12 @@
+package com.avab.avab.service;
+
+import com.avab.avab.dto.response.AuthResponseDTO.OAuthResponse;
+import com.avab.avab.dto.response.AuthResponseDTO.TokenRefreshResponse;
+
+public interface AuthService {
+
+    OAuthResponse kakaoLogin(String code);
+
+    TokenRefreshResponse refresh(String refreshToken);
+
+}
